Geo3o1_6 (29.09.2022)
In the previous version, the plotting was done by the errors estimated disparities from the points imported to plotting function.
In this new version, the plotting function directly imports the disparities where they are estimated at first. The disparities are assigned to the base w.r.t. the direct georeferencing, pre-and bundle adjustment, and also considering the stereo or triplet case.